Fix entity lookup and pass 1.4k more tests
Our entity matching is still not 100% correct, but this at least fixes the bug where it was forgetting which entity range had already matched the prefix by not updating firstPossible, meaning it would return the first entity which matched a character at the end of the entity.

This still has a bug of not matching the longest possible entity so many tests have an extra ";" as a result.