Add super basic treebuilding
The problem was "break reprocess" didn't jump backwards like I thought it would.  Still not sure what it does.  Instead I'm using a while loop and "continue" which is also confusing, but at least works.

Also added Document to Node and pasted in a bunch more of the spec as comments.